refactor: split ai chat and assignment sections
