+++ docs/dev-logs/2026-06-06-frontend-assignment-ai-chat-split-bundle-v7.md
@@ -1,0 +1,13 @@
+# Frontend Assignment and AI Chat Split Bundle v7
+
+## Summary
+- Split `AIChatPageSections` into dedicated hero and conversation sections.
+- Split `AdminAssignPageSections` into hero, instructor, and student sections while preserving the original page API.
+
+## Verification
+- `npm run build:frontend`
+- `npm run verify`
+
+## Notes
+- Frontend build passed after the split.
+- Backend test suite passed unchanged at 74 tests.
